+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+#### v 1.0.0 - 2019.MM.DD
+* Now it uses threading services of py-serial instead of py-serial-asyncio
+* Introduced the possibility of set one or more reserved keys by the user in Arancino Library with a new command: `CMD_APP_SET_RSVD`
+
 #### v 0.1.5 - 2019.07.23
 * Fix Arancino Service